+++ R/FindOptimalSubset.R
@@ -117,9 +117,9 @@
 #' }
 #'
 
-FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100L,
-                              migrationRate=0.1, migrationInterval=10L,
-                              pcrossover=0.8, pmutation=0.1, elitism=0L,
+FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100,
+                              migrationRate=0.1, migrationInterval=10,
+                              pcrossover=0.8, pmutation=0.1, elitism=0,
                               maxiter=1000L, run=maxiter, suggestions=NULL,
                               parallel=TRUE, seed=NULL) {
 
@@ -159,8 +159,10 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100L,
         set.seed(seed); m <- t(apply(m, 1, sample, size=k))
       } else if (k > ncol(m)) {
         idxs <- seq_len(n)
-        FUN <- function(i) c(i, sample(idxs[-i], k - ncol(m)))
-        set.seed(seed); m <- t(apply(m, 1, FUN))
+        set.seed(seed)
+        m <- t(apply(m, 1, function(i) {
+          c(i, sample(idxs[-i], k - ncol(m)))
+        }))
       }
       suggestions <- t(apply(m, 1, function(i) EncodeChromosome(i, n)))
     }
@@ -197,8 +199,9 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100L,
   })
 
   # decode solution
-  FUN <- function(i) sort(DecodeChromosome(i, n))
-  m <- t(apply(ga_output@solution, 1, FUN))
+  m <- t(apply(ga_output@solution, 1, function(i) {
+    sort(DecodeChromosome(i, n))
+  }))
   solution <- m[!duplicated(m), , drop=FALSE]
 
   # bundle output
@@ -237,8 +240,9 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100L,
 .Crossover <- function(object, parents, n) {
   fitness_parents <- object@fitness[parents]
   encoded_parents <- object@population[parents, , drop=FALSE]
-  FUN <- function(i) DecodeChromosome(i, n)
-  decoded_parents <- t(apply(encoded_parents, 1, FUN))
+  decoded_parents <- t(apply(encoded_parents, 1, function(i) {
+    DecodeChromosome(i, n)
+  }))
   p1 <- decoded_parents[1, ]
   p2 <- decoded_parents[2, ]
   c1 <- p1
@@ -249,8 +253,9 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100L,
   c1[i1] <- p2[i1]
   c2[i2] <- p1[i2]
   decoded_children <- rbind(c1, c2)
-  FUN <- function(i) EncodeChromosome(i, n)
-  encoded_children <- t(apply(decoded_children, 1, FUN))
+  encoded_children <- t(apply(decoded_children, 1, function(i) {
+    EncodeChromosome(i, n)
+  }))
   m <- t(apply(object@population, 1, function(i) sort(DecodeChromosome(i, n))))
   FindFitness <- function(child) {
     return(object@fitness[which(apply(m, 1, function(i) identical(i, child)))[1]])
@@ -297,8 +302,9 @@ FindOptimalSubset <- function(n, k, Fitness, ..., popSize=100L,
 
 EncodeChromosome <- function(x, n) {
   width <- ceiling(log2(n + 1))
-  FUN <- function(i) GA::decimal2binary(i, width)
-  return(unlist(lapply(x, FUN)))
+  return(unlist(lapply(x, function(i) {
+    GA::decimal2binary(i, width)
+  })))
 }
 
 #' @rdname EncodeChromosome
@@ -306,6 +312,7 @@ EncodeChromosome <- function(x, n) {
 
 DecodeChromosome <- function(y, n) {
   width <- ceiling(log2(n + 1))
-  FUN <- function(i) GA::binary2decimal(y[i:(i + width - 1L)])
-  return(vapply(seq(1, length(y), by=width), FUN, 0))
+  return(vapply(seq(1, length(y), by=width), function(i) {
+    GA::binary2decimal(y[i:(i + width - 1L)])
+  }, 0))
 }